Move incremental_walk -> EntityPath
