+++ crates/re_log_types/src/path/entity_path.rs
@@ -184,6 +184,23 @@ impl EntityPath {
     pub fn join(&self, other: &Self) -> Self {
         self.iter().chain(other.iter()).cloned().collect()
     }
+
+    /// Helper function to iterate over all incremental [`EntityPath`]s from start to end, NOT including start itself.
+    ///
+    /// For example `incremental_walk("foo", "foo/bar/baz")` returns: `["foo/bar", "foo/bar/baz"]`
+    pub fn incremental_walk<'a>(
+        start: Option<&'_ EntityPath>,
+        end: &'a EntityPath,
+    ) -> impl Iterator<Item = EntityPath> + 'a {
+        re_tracing::profile_function!();
+        if start.map_or(true, |start| end.is_descendant_of(start)) {
+            let first_ind = start.map_or(0, |start| start.len() + 1);
+            let parts = end.as_slice();
+            itertools::Either::Left((first_ind..=end.len()).map(|i| EntityPath::from(&parts[0..i])))
+        } else {
+            itertools::Either::Right(std::iter::empty())
+        }
+    }
 }
 
 impl SizeBytes for EntityPath {
@@ -360,5 +377,48 @@ impl std::fmt::Debug for EntityPath {
 impl std::fmt::Display for EntityPath {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         self.path.fmt(f)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_incremental_walk() {
+        assert_eq!(
+            EntityPath::incremental_walk(None, &EntityPath::root()).collect::<Vec<_>>(),
+            vec![EntityPath::root()]
+        );
+        assert_eq!(
+            EntityPath::incremental_walk(Some(&EntityPath::root()), &EntityPath::root())
+                .collect::<Vec<_>>(),
+            vec![]
+        );
+        assert_eq!(
+            EntityPath::incremental_walk(None, &EntityPath::from("foo")).collect::<Vec<_>>(),
+            vec![EntityPath::root(), EntityPath::from("foo")]
+        );
+        assert_eq!(
+            EntityPath::incremental_walk(Some(&EntityPath::root()), &EntityPath::from("foo"))
+                .collect::<Vec<_>>(),
+            vec![EntityPath::from("foo")]
+        );
+        assert_eq!(
+            EntityPath::incremental_walk(None, &EntityPath::from("foo/bar")).collect::<Vec<_>>(),
+            vec![
+                EntityPath::root(),
+                EntityPath::from("foo"),
+                EntityPath::from("foo/bar")
+            ]
+        );
+        assert_eq!(
+            EntityPath::incremental_walk(
+                Some(&EntityPath::from("foo")),
+                &EntityPath::from("foo/bar/baz")
+            )
+            .collect::<Vec<_>>(),
+            vec![EntityPath::from("foo/bar"), EntityPath::from("foo/bar/baz")]
+        );
     }
 }
